Remove orphan Farms
The relationship between Farm and Project is One-To-One. One-To-One relations are
asymmetrical. This means that one object will always depend on the other. In our Promgen,
a Farm will always be linked to a specific Project. We have also removed the feature
that links a Project with an existing Farm in the database so that no one will be able
to use the existing orphan Farms, and these Farms will become redundant. Therefore, we
have created a migration script to clean up these orphan Farms.