Always use historical models in migrations.RunPython functions
When running migrations, Django operates from historical versions of models stored in
migration files. Therefore, if we write migrations using the RunPython operation, we
need to use these historical model versions rather than importing them directly. If
we import models directly instead of using the historical models, migrations may work
initially but will fail in the future.[1]

As a result, we have fixed all direct imports of Promgen's models that existed in the
current migration files to avoid errors in the future.

1: https://docs.djangoproject.com/en/5.2/topics/migrations/#historical-models