Prevent users from accessing the default group
Promgen already had a group object named "Default group" before the Membership feature
was introduced. This group is being used for managing permissions in the current
authorization model of Promgen. We have excluded this group from the new features
related to the Membership feature, which is only about user's group management.