Remove excluded "farm" field from ProjectSerializer
After removing "farm" field from the Project model, calling the Project Retrieve API
will cause an AssertionError. This happened because the serializer cannot find the
"farm" field which was declared in the excluded field list. We have removed the "farm"
field from the excluded field list of ProjectSerializer to resolve this bug.