Add a self-test for checking the deletion of a project without a farm
In the past, an error occurred when we forgot to check for the existence of a farm
before attempting to access it at the time the delete_project signal was triggered.
This caused the project deletion action to throw a RelatedObjectDoesNotExist exception
when it tried to access the farm of the project being deleted. As a result, the deletion
of the project or its parent service failed, and Promgen Web displayed an Internal Error.

We have added a self-test for this case to ensure that the bug has been fixed and
will always be checked in the future.