Fix the bug that the project's owner lost admin permission
In Promgen, there's currently an issue where a project's owner might not have admin
permission on that project. This happens when a user, A, is both the owner of a project
and has permission on its parent service. If another user, B, tries to remove user A's
permission from the service and checks the "Remove permissions on child projects" option,
user A will lose admin permission on the project, even though they remain the owner.

This situation goes against our Promgen rule, which ensures that an object's owner always
has admin permission on it. However, actions on a service should have a greater impact
than on its child projects. So, we've decided to respect user B's action in this case.
Therefore, to resolve this issue, the removal of permissions will proceed successfully,
and the project's ownership will be transferred to the owner of the parent service.
A warning also was added to notice user about the change of the owner.