Add a migration file for re-assigning the owner for projects
At commit 7222e82e, we fixed the bug where the project's owner lost admin permission.
We added a migration file to re-assign the owner to projects which have this data
inconsistency. The ownership of the project is transferred from the current owner to the
service's owner, and also the admin permission is granted to these new owners.