Add Audit Logs when assigning or removing permission for a Group
We introduced a feature that allows users to assign or remove permission for a Group.
Recording these actions as Audit Logs would be useful for users. Therefore, we have
linked the GroupObjectPermission model to the create_log and delete_log functions using
the post_save and post_delete signals to generate audit logs when permissions are assigned
to or removed from a group.
However, the post_delete signal will be temporarily disconnected when migration file
number 0033 is executed. In this scenario, Promgen removes permissions automatically, so
Audit Logs are not needed.