Add pre_save signal for cleaning existing permissions
In Promgen, we always want to ensure that a user has at most one permission on a specific
object. We achieved this by removing all existing permissions of a user on an object before
any new permission assignment. This solution is not optimal because it creates duplicate
codes and could also be overlooked.
Therefore, we have replaced these codes with a new signal that is responsible for removing
all existing permissions just before the new permission is saved to the database.