Fix the bug that links in the audit log table may not work
Currently, on Promgen's Edit History page, the links in the audit log table sometimes do
not function as expected. There are several reasons for this:

- The get_absolute_url function has not been implemented for the model corresponding to
the log entry.
- The model corresponding to the log entry is not managed by Promgen.
- The object corresponding to the log entry has been deleted from the database.

To improve this feature, we have created a new template tag to support link generation.
This allows for reduced dependency on model classes and makes link generation more
flexible. For example, in cases where the object has been deleted, we will create a link
that navigates to the parent object of that object.