Fix bug that notifier stops working after changing username
We resolved an issue where notifiers of the "user" type would stop working if a username was
changed. This problem occurred because the notifier's value was initially set to the username
at the time of creation. Consequently, any changes to the username were not reflected in the
notifier's value, causing it to lose its association with the User object.

To address this, we now use the user's primary key as the notifier's value instead of the
username. Since the PK is never changed, the notifier will consistently link to the correct
User object. A migration script has been also added to migrate the existing data.