Add check per-object permissions for ExporterScrape view
Currently, a Django View class is used for the Exporter Scrape feature, but it returns
JSON instead of HTML. This makes it difficult to use PromgenGuardianPermissionMixin for
permission checks because this View class is used as an API rather than a typical View
class. However, if we don't check permissions on ExporterScrape, a security vulnerability
could arise in the future, allowing a user to view the list of hosts for a project they
don't have access to. Therefore, it is necessary to check permissions on ExporterScrape.

The ideal solution for permission checking would be to replace this View class with a true
REST API and implement permission checks for the API. However, this solution requires a
significant amount of code changes. We also plan to rewrite a completely new version of
Promgen's APIs soon. Therefore, to address the permission check issue on ExporterScrape,
we have chosen a simpler solution by manually checking permissions directly in the "post"
function of this class.