Improve the Show Alert's Raw Data Feature
Currently, users have to click the Raw Data text on the Alert Detail page to see the JSON
content of the alert. Using plain text makes it unclear that it is clickable. Even when
users click it, there is a rendering bug that prevents the screen from scrolling down
automatically, causing users to think this feature doesn't work.

We have resolved these issues by rewriting this feature. The <a> HTML tag with an additional
icon has been used, and the collapsible area has been replaced by a dialog modal to improve
the user experience.