Add Merge Users feature
We have introduced a new feature to the User model on the Administrator site that allows Promgen's
admins to merge a user into another. After the merge, the kept user will assume ownership of all
associated objects. User roles will be consolidated, retaining the highest level of permissions for
each object. The old user will be permanently removed.