Fix bug that registering an invalid URL causes an internal error
We have fixed a bug that caused an internal server error when registering a URL with an invalid
value. The root cause was that we removed a file named url_form.html at commit acb124f3.
Although Django's URLField already has URL validation, Django needs a template to display error
messages. Therefore, Django encountered an error when trying to find the removed template.

Restoring the template file is not a good solution because it will redirect users to another
page when the form is invalid, which diminishes user experience. We fixed the bug by displaying
the error message at the top of the Project Detail page instead of rendering a new interface
based on the template file.